# docs/swagger.yaml
basePath: /api/v1
definitions:
  database.Ability:
    properties:
      name:
        type: string
      text:
        type: string
    type: object
  database.Agenda:
    properties:
      name:
        type: string
      text:
        type: string
    type: object
  database.Civilization:
    properties:
      ability:
        $ref: '#/definitions/database.Ability'
      infrastructure:
        type: string
      leaders:
        items:
          type: string
        type: array
      name:
        type: string
      unit:
        type: string
    type: object
  database.District:
    properties:
      buildings:
        items:
          type: string
        type: array
      description:
        type: string
      exclusive_to:
        type: string
      name:
        type: string
      plunder_yields:
        type: string
    type: object
  database.Improvement:
    properties:
      description:
        type: string
      name:
        type: string
      placement:
        items:
          type: string
        type: array
      plunder:
        type: string
      resources:
        items:
          type: string
        type: array
      technology:
        type: string
    type: object
  database.Leader:
    properties:
      ability:
        $ref: '#/definitions/database.Ability'
      agenda:
        $ref: '#/definitions/database.Agenda'
      civ:
        type: string
      name:
        type: string
      unit:
        type: string
    type: object
host: localhost:8000
info:
  contact:
    name: grqphical
    url: https://github.com/grqphical07
  description: A free-to-use API to access information about Sid Meier's Civilization
    VI
  license:
    name: MIT License
    url: https://github.com/grqphical07/rulers-compendium/blob/main/LICENSE
  title: Rulers Compendium API
  version: "1.0"
paths:
  /api/v1/civilizations:
    get:
      consumes:
      - '*/*'
      description: Get's all civilizations available to play in civ 6
      parameters:
      - description: limits amount of results returned
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Civilization'
            type: array
        "400":
          description: Invalid limit value
          schema:
            type: string
      tags:
      - Civilizations
  /api/v1/civilizations/{name}:
    get:
      consumes:
      - '*/*'
      description: Gets a civilization in civ 6 based on a given name
      parameters:
      - description: civilization to get
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Civilization'
      tags:
      - Civilizations
  /api/v1/districts:
    get:
      consumes:
      - '*/*'
      description: Get's all districts from civ 6
      parameters:
      - description: limits amount of results returned
        in: query
        name: limit
        type: integer
      - description: filters leaders by civilization
        in: query
        name: civilization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.District'
            type: array
        "400":
          description: Invalid limit value
          schema:
            type: string
      tags:
      - Districts
  /api/v1/districts/{name}:
    get:
      consumes:
      - '*/*'
      description: Gets a district by name in civ 6
      parameters:
      - description: district to find
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.District'
      tags:
      - Districts
  /api/v1/improvements:
    get:
      consumes:
      - '*/*'
      description: Get's all improvements buildable in civ 6
      parameters:
      - description: limits amount of results returned
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Improvement'
            type: array
        "400":
          description: Invalid limit value
          schema:
            type: string
      tags:
      - Improvements
  /api/v1/improvements/{name}:
    get:
      consumes:
      - '*/*'
      description: Gets an improvement in civ 6 based on a given name
      parameters:
      - description: Improvement to get
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Improvement'
      tags:
      - Improvements
  /api/v1/leaders:
    get:
      consumes:
      - '*/*'
      description: Get's all leaders from civ 6 including their agenda's and abilities
      parameters:
      - description: limits amount of results returned
        in: query
        name: limit
        type: integer
      - description: filters leaders by civilization
        in: query
        name: civilization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Leader'
            type: array
        "400":
          description: Invalid limit value
          schema:
            type: string
      tags:
      - Leaders
  /api/v1/leaders/{name}:
    get:
      consumes:
      - '*/*'
      description: Gets a single leader from civ 6
      parameters:
      - description: leader to get
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Leader'
      tags:
      - Leaders
schemes:
- http
swagger: "2.0"